refactor connection layout
